+++ Runtime/AssemblyInfo.cs
@@ -1,3 +1,11 @@
+using System.ComponentModel;
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("At.Ac.FhStp.Import3D.Tests")]
+
+// ReSharper disable once CheckNamespace
+namespace System.Runtime.CompilerServices
+{
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    internal class IsExternalInit{}
+}